services/blog-service/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
python-jose[cryptography]==3.3.0
httpx==0.27.0
sqlalchemy==2.0.30
asyncpg==0.29.0
pydantic-settings==2.2.1